Isolated router and recording code
